requirements.txt
azure-storage-blob<13
azure-storage-file-datalake
duckdb<2
polars<2
pyodbc<6
python-dotenv<2
sqlalchemy<3
tqdm<5